+++ backend/src/graphql/mutations/comment.resolver.ts
@@ -1,5 +1,5 @@
 import { MutationResolvers } from 'src/graphql/generated/graphql'
-import { connection } from '../../mysql/connect'
+import { connection } from '../../database/mysql'
 
 const sql = 'insert into comment (content) values ($1)' // mysql도 이거 맞나?
 
@@ -26,7 +26,7 @@ export const Mutation: MutationResolvers = {
     })
   },
 
-  deleteComment: () => {
+  deleteComment: (_, { id }) => {
     return new Promise((resolve, reject) => {
       connection.query(sql, (err: Error, row: any, cols: any) => {
         if (err) {
