+++ backend/src/graphql/queries/user.resolver.ts
@@ -1,5 +1,5 @@
 import { QueryResolvers } from 'src/graphql/generated/graphql'
-import { connection } from '../../mysql/connect'
+import { connection } from '../../database/mysql'
 
 const sql = 'select * from users;'
